c++: Fix importing nested namespace declarations [PR100707]
Bootstrapped and regtested on x86_64-pc-linux-gnu, OK for trunk?

-- >8 --

Currently, importing a namespace declarations marks it as imported, and
so marks it as originating from the module that it was imported from.
This is usually harmless, but causes problems with nested namespaces.

In the linked PR, what happens is that the namespace 'A' imported from
the module ends up not being considered when creating the 'A' namespace
within its own TU, and thus it has its 'cp_binding_level' recreated.
However, by this point 'A::B' has already been imported, and so the
'level_chain' member no longer correctly points at 'A's binding level,
so the sanity check for this in 'resume_scope' ICEs.

Since as far as I can tell there's no reason for imported namespaces to
be attached to any specific module (namespace declarations with external
linkage are always attached to the global module by [module.unit] p7.2),
this patch just removes the 'imported' flag, which stops code from
caring about its originating module.

This patch also makes some minor adjustments to existing tests to cater
for the new dumped name.

	PR c++/100707

gcc/cp/ChangeLog:

	* name-lookup.cc (add_imported_namespace): Don't mark namespaces
	as imported.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/indirect-1_b.C: Adjust to handle namespaces not
	being attached to the module they were imported from.
	* g++.dg/modules/indirect-1_c.C: Likewise.
	* g++.dg/modules/indirect-2_b.C: Likewise.
	* g++.dg/modules/indirect-2_c.C: Likewise.
	* g++.dg/modules/indirect-3_b.C: Likewise.
	* g++.dg/modules/indirect-3_c.C: Likewise.
	* g++.dg/modules/indirect-4_b.C: Likewise.
	* g++.dg/modules/indirect-4_c.C: Likewise.
	* g++.dg/modules/namespace-5_a.C: New test.
	* g++.dg/modules/namespace-5_b.C: New test.
	* g++.dg/modules/namespace-5_c.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>