Adjust expectations for pr59533-1.c
The change for pr111267 twiddled code generation for sh/pr59533-1.c

We end up eliminating two comparisons, but require two shll instructions
to do so.  And in a couple places we're using an addc sequence rather
than a subc sequence.   This patch adjusts the expected codegen for the
test as all those are either a wash or a

The fwprop change does cause some code regressions on the same test.
I'll file a distinct but for that issue.

Pushed to the trunk,

Jeff
commit 7e16f819ff413c48702f9087b62eaac39a060a14
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Sun Jan 21 19:12:21 2024 -0700

    [committed] Adjust expectations for pr59533-1.c

    The change for pr111267 twiddled code generation for sh/pr59533-1.c

    We end up eliminating two comparisons, but require two shll instructions to do
    so.  And in a couple places we're using an addc sequence rather than a subc
    sequence.   This patch adjusts the expected codegen for the test as all those
    are either a wash or a

    The fwprop change does cause some code regressions on the same test.  I'll file
    a distinct but for that issue.

    gcc/testsuite
            * gcc.target/sh/pr59533-1.c: Adjust expected output.