c++/modules: Handle error header names in modules [PR107594]
I don't provide a new test because this error only happens when there
are no include paths at all, and I haven't worked out a way to get this
to happen within DejaGNU (as it adds a number of `-B` and `-I` flags).

Bootstrapped and regtested on x86_64-pc-linux-gnu, OK for trunk?

-- >8 --

When there are no include paths while preprocessing a header-name token,
an empty STRING_CST is returned. This patch ensures this is handled when
attempting to create a module for this name.

	PR c++/107594

gcc/cp/ChangeLog:

	* module.cc (get_module): Bail on empty name.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>