PR111022 ES0.0E0 format gave ES0.dE0 output with d too high.
The attached patch attempts to fix the handling of the EN0.0E0 and
ES0.0E0 formatting by correctly calculating the number of digits needed
for the exponents and building those exponents into the float string.

My editor judiciously deleted trailing blank spaces in a number of
places.  I apologize for the clutter, but we might as well get rid of it
now.

Two existing test cases needed to be adjusted and I am adding one new
test case to capture the changes in our testsuite.

Regression tested on X86_64.

OK for trunk?  Do we need to backport this?

Regards,

Jerry

Author: Jerry DeLisle <jvdelisle@gcc.gnu.org>
Date:   Tue Jan 30 09:45:49 2024 -0800

     libgfortran: EN0.0E0 and ES0.0E0 format editing.

             PR libgfortran/111022

     F2018 and F2023 standards added zero width exponents. This required
     additional special handing in the process of building formatted
     floating point strings.

     libgfortran/ChangeLog:

             * io/write.c (select_buffer): Whitespace.
             (write_real): Whitespace.
             (write_real_w0): Adjust logic for d==0.
             * io/write_float.def (determine_precision): Whitespace.
             (build_float_string): Calculate the width of the E0 exponents.
             (build_infnan_string): Whitespace.
             (CALCULATE_EXP): Whitespace
             (quadmath_snprintf): Whitespace.
             (determine_en_precision): Whitespace.

     gcc/testsuite/ChangeLog:

             * gfortran.dg/pr96436_4.f90: Changed for ES0 and EN0.
             * gfortran.dg/pr96436_5.f90: Changed for ES0 and EN0.
             * gfortran.dg/pr111022.f90: New test.